Fix undefined behavior in kernel shutdown
Previously we were shutting the kernel down with sys.exit(). Exiting
with sys.exit() when the process was created after a fork may trigger
undefined behavior, since cleanup functions run twice, once for the
parent process and once for the child process.

This change uses os._exit() instead to kill the child process, as
recommended by Python docs.

This change also makes the SIGTERM handler re-entrant, to fix a bug
in which the virtual file registry shutdown method was being interrupted
upon receipt of a second SIGTERM. We terminate the process once
manually, and because the process is a daemon process it gets another
SIGTERM.

+++ marimo/_runtime/virtual_file.py
@@ -98,6 +98,7 @@ class VirtualFileRegistry:
     registry: dict[str, shared_memory.SharedMemory] = dataclasses.field(
         default_factory=dict
     )
+    shutting_down = False
 
     def __del__(self) -> None:
         self.shutdown()
@@ -145,9 +146,19 @@ class VirtualFileRegistry:
             del self.registry[key]
 
     def shutdown(self) -> None:
-        for _, shm in self.registry.items():
-            shm.unlink()
-        self.registry.clear()
+        # Try to make this method re-entrant since it's called in the
+        # sigterm handler
+        #
+        # https://www.gnu.org/software/libc/manual/html_node/Nonreentrancy.html
+        if self.shutting_down:
+            return
+        try:
+            self.shutting_down = True
+            for _, shm in self.registry.items():
+                shm.unlink()
+            self.registry.clear()
+        finally:
+            self.shutting_down = False
 
 
 def _without_leading_dot(ext: str) -> str:
